rita top

+++ backend/backend/Dto/ApprovedMessageResponse.cs
@@ -3,8 +3,8 @@
 public class ApprovedMessageResponse
 {
     public string Title { get; set; }
-    public int ChatId { get; set; }
-    public int MessageId { get; set; }
+    public long ChatId { get; set; }
+    public long MessageId { get; set; }
     public string SessionString { get; set; }
     public string Content { get; set; }
 }
