+++ backend/backend/Entities/ApprovedMessage.cs
@@ -6,8 +6,8 @@ public class ApprovedMessage
     public string Keyword { get; set; }
     public string Title { get; set; }
     public bool IsFound { get; set; }
-    public int ChatId { get; set; }
-    public int MessageId { get; set; }
+    public long ChatId { get; set; }
+    public long MessageId { get; set; }
     public string SessionString { get; set; }
     public string Content { get; set; }
 }
